Fix volume truncation bug in GoldTrendBot and VolumeProfileMtfBot
CalculatePositionVolume returned long and cast the normalized volume
to it, silently truncating any fractional volume below 1 unit to 0.
This only stays hidden on instruments whose minimum volume step is a
whole unit (e.g. XAUUSD at 1 oz per 0.01 lot on Pepperstone); on an
instrument with a fractional step (e.g. a crypto CFD), a validly
risk-sized position would be truncated to 0 and every entry would be
silently skipped regardless of risk settings. ExecuteMarketOrder takes
a double volume, so the long return type and cast were both
unnecessary. Found while reviewing a new bot that shares the same
sizing pattern.

+++ GoldTrendBot/GoldTrendBot/GoldTrendBot/GoldTrendBot.cs
@@ -171,7 +171,7 @@ namespace cAlgo.Robots
                 Print("Order failed: {0}", result.Error);
         }
 
-        private long CalculatePositionVolume(double stopLossPips)
+        private double CalculatePositionVolume(double stopLossPips)
         {
             var riskAmount = Account.Balance * (RiskPercent / 100.0);
 
@@ -186,7 +186,7 @@ namespace cAlgo.Robots
             if (normalized > Symbol.VolumeInUnitsMax)
                 normalized = Symbol.VolumeInUnitsMax;
 
-            return (long)normalized;
+            return normalized;
         }
 
         private void ManageTrailingStop()
